requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
Pillow>=8.0.0
tqdm>=4.50.0
matplotlib>=3.3.0
scikit-learn>=0.24.0  # For KFold cross-validation
networkx>=2.5  # Optional, for visualization